vector-store: split `db_cdc` into a module
The actor loop and the CDC consumer are two fairly independent
responsibilities that happened to share one large file.
Splitting them into their own submodules keeps each one focused
and makes room for adding more CDC-related code without the file
growing without bound.
